add the FPGA module back
