+++ Makefile
@@ -104,7 +104,6 @@ SRC_C += mphalport.c
 
 SRC_C += driver/bluetooth_data_protocol.c
 SRC_C += driver/bluetooth_low_energy.c
-SRC_C += driver/ecx336cn.c
 SRC_C += driver/flash.c
 SRC_C += driver/fpga.c
 SRC_C += driver/spi.c
@@ -119,7 +118,7 @@ SRC_C += micropython/extmod/modure.c
 #SRC_C += modules/camera.c
 SRC_C += modules/device.c
 #SRC_C += modules/display.c
-#SRC_C += modules/fpga.c
+SRC_C += modules/fpga.c
 SRC_C += modules/led.c
 SRC_C += modules/time.c
 SRC_C += modules/touch.c
